+++ apps/meteor/app/lib/server/lib/sendNotificationsOnMessage.js
@@ -1,7 +1,5 @@
-import { Room } from '@rocket.chat/core-services';
 import { Subscriptions, Users } from '@rocket.chat/models';
 import emojione from 'emojione';
-import { Meteor } from 'meteor/meteor';
 import moment from 'moment';
 
 import { callbacks } from '../../../../lib/callbacks';
@@ -49,12 +47,13 @@ export const sendNotification = async ({
 		subscription.receiver = [
 			await Users.findOneById(subscription.u._id, {
 				projection: {
-					active: 1,
-					emails: 1,
-					language: 1,
-					status: 1,
-					statusConnection: 1,
-					username: 1,
+					'active': 1,
+					'emails': 1,
+					'language': 1,
+					'status': 1,
+					'statusConnection': 1,
+					'username': 1,
+					'settings.preferences.enableMobileRinging': 1,
 				},
 			}),
 		];
@@ -69,6 +68,7 @@ export const sendNotification = async ({
 	}
 
 	const isThread = !!message.tmid && !message.tshow;
+	const isVideoConf = message.t === 'videoconf';
 
 	notificationMessage = await parseMessageTextPerUser(notificationMessage, message, receiver);
 
@@ -113,6 +113,9 @@ export const sendNotification = async ({
 			hasReplyToThread,
 			roomType,
 			isThread,
+			isVideoConf,
+			userPreferences: receiver.settings?.preferences,
+			roomUids: room.uids,
 		})
 	) {
 		queueItems.push({
@@ -201,6 +204,7 @@ const project = {
 		'receiver.status': 1,
 		'receiver.statusConnection': 1,
 		'receiver.username': 1,
+		'receiver.settings.preferences.enableMobileRinging': 1,
 	},
 };
 
@@ -353,50 +357,7 @@ export async function sendAllNotifications(message, room) {
 		return message;
 	}
 
-	const { sender, hasMentionToAll, hasMentionToHere, notificationMessage, mentionIds, mentionIdsWithoutGroups } =
-		await sendMessageNotifications(message, room);
-
-	// on public channels, if a mentioned user is not member of the channel yet, he will first join the channel and then be notified based on his preferences.
-	if (room.t === 'c') {
-		// get subscriptions from users already in room (to not send them a notification)
-		const mentions = [...mentionIdsWithoutGroups];
-		const cursor = Subscriptions.findByRoomIdAndUserIds(room._id, mentionIdsWithoutGroups, {
-			projection: { 'u._id': 1 },
-		});
-
-		for await (const subscription of cursor) {
-			const index = mentions.indexOf(subscription.u._id);
-			if (index !== -1) {
-				mentions.splice(index, 1);
-			}
-		}
-
-		const users = await Promise.all(
-			mentions.map(async (userId) => {
-				await Room.join({ room, user: { _id: userId } });
-
-				return userId;
-			}),
-		).catch((error) => {
-			throw new Meteor.Error(error);
-		});
-
-		const subscriptions = await Subscriptions.findByRoomIdAndUserIds(room._id, users).toArray();
-		users.forEach((userId) => {
-			const subscription = subscriptions.find((subscription) => subscription.u._id === userId);
-
-			void sendNotification({
-				subscription,
-				sender,
-				hasMentionToAll,
-				hasMentionToHere,
-				message,
-				notificationMessage,
-				room,
-				mentionIds,
-			});
-		});
-	}
+	await sendMessageNotifications(message, room);
 
 	return message;
 }