+++ apps/meteor/client/views/room/composer/messageBox/MessageBoxActionsToolbar/hooks/useCreateDiscussionAction.tsx
@@ -1,0 +1,33 @@
+import type { IRoom } from '@rocket.chat/core-typings';
+import { isRoomFederated } from '@rocket.chat/core-typings';
+import { useTranslation, useSetting, usePermission, useSetModal } from '@rocket.chat/ui-contexts';
+import React from 'react';
+
+import CreateDiscussion from '../../../../../../components/CreateDiscussion';
+import type { GenericMenuItemProps } from '../../../../../../components/GenericMenu/GenericMenuItem';
+
+export const useCreateDiscussionAction = (room?: IRoom): GenericMenuItemProps => {
+	const t = useTranslation();
+	const setModal = useSetModal();
+
+	if (!room) {
+		throw new Error('Invalid room');
+	}
+
+	const handleCreateDiscussion = () =>
+		setModal(<CreateDiscussion onClose={() => setModal(null)} defaultParentRoom={room?.prid || room?._id} />);
+
+	const discussionEnabled = useSetting('Discussion_enabled') as boolean;
+	const canStartDiscussion = usePermission('start-discussion', room._id);
+	const canSstartDiscussionOtherUser = usePermission('start-discussion-other-user', room._id);
+
+	const allowDiscussion = room && discussionEnabled && !isRoomFederated(room) && (canStartDiscussion || canSstartDiscussionOtherUser);
+
+	return {
+		id: 'create-discussion',
+		content: t('Discussion'),
+		icon: 'discussion',
+		disabled: !allowDiscussion,
+		onClick: handleCreateDiscussion,
+	};
+};