+++ apps/meteor/client/views/room/contextualBar/Threads/components/ThreadChat.tsx
@@ -110,17 +110,17 @@ const ThreadChat = ({ mainMessage }: ThreadChatProps) => {
 						onUploadFiles={handleUploadFiles}
 						tshow={sendToChannel}
 					>
-						<Field>
-							<FieldRow marginBlock={8}>
-								<FieldLabel htmlFor={sendToChannelID} color='annotation' fontScale='p2'>
-									{t('Also_send_to_channel')}
-								</FieldLabel>
+						<Field marginBlock={8}>
+							<FieldRow justifyContent='initial'>
 								<CheckBox
 									id={sendToChannelID}
 									checked={sendToChannel}
 									onChange={() => setSendToChannel((checked) => !checked)}
 									name='alsoSendThreadToChannel'
 								/>
+								<FieldLabel mis='x8' htmlFor={sendToChannelID} color='annotation' fontScale='p2'>
+									{t('Also_send_to_channel')}
+								</FieldLabel>
 							</FieldRow>
 						</Field>
 					</ComposerContainer>