+++ apps/meteor/server/services/meteor/service.ts
@@ -152,9 +152,11 @@ export class MeteorService extends ServiceClassInternal implements IMeteor {
 					return;
 				}
 
-				serviceConfigCallbacks.forEach((callbacks) => {
-					callbacks[clientAction === 'inserted' ? 'added' : 'changed']?.(id, data);
-				});
+				if (data) {
+					serviceConfigCallbacks.forEach((callbacks) => {
+						callbacks[clientAction === 'inserted' ? 'added' : 'changed']?.(id, data);
+					});
+				}
 			});
 		}
 
@@ -222,8 +224,8 @@ export class MeteorService extends ServiceClassInternal implements IMeteor {
 
 		if (!disableMsgRoundtripTracking) {
 			listenToMessageSentEvent(this, async (message) => {
-				if (message?._updatedAt) {
-					metrics.messageRoundtripTime.set(Date.now() - message._updatedAt.getDate());
+				if (message?._updatedAt instanceof Date) {
+					metrics.messageRoundtripTime.observe(Date.now() - message._updatedAt.getTime());
 				}
 			});
 		}